Extracted solicitor DR management into new spec
